requirements.txt
# LLM Integration
openai>=1.0.0                # genai_studio.py — OpenAI-compatible client
httpx>=0.25.0                # genai_studio.py — HTTP transport for OpenAI client

# Configuration
PyYAML>=6.0                  # src/config.py — position profile YAML loading

# CLI
click>=8.0.0                 # src/cli.py — command-line interface
rich>=13.0.0                 # src/cli.py — progress bars, tables, console output

# Web Viewer
flask>=3.0.0                 # src/viewer/app.py — applicant review dashboard

# Document Processing (optional — graceful fallback if missing)
pdfplumber>=0.10.0           # src/extraction/pdf_extractor.py — primary PDF extractor
pypdf>=4.0.0                 # src/extraction/pdf_extractor.py — fallback PDF extractor
python-docx>=1.0.0           # src/extraction/docx_extractor.py — .docx text extraction

# Testing (dev only)
# pytest>=7.0.0